requirements.txt
Flask
Jinja2
python-dotenv
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-User